hirni tests to use a local clone of toolbox
The default toolbox to be used by hirni doesn't need
to be cloned (and get the conatainers) over network
over and over again. Use a single local clone instead.

This used to get us into trouble, flooding the
original data source leading to failure to actually
get the conatainers. While the denial on the remote
end is something to solve itself, even w/o that issue
it seems reasonable and faster to have that locally.