ENH: test-datalad: Enable SSH testing
Make use of the Docker SSH target that DataLad supports as of
datalad/datalad#4762.

Note that DataLad's master version of prep-travis-forssh.sh is used
for maint as well.  That's simple and should work fine for maint and
the latest release (once a release contains the changes from PR 4762).
But if this approach ends up causing problems, we can of course switch
to a more specific download or to cloning the repo and checking out
the appropriate revision.

Closes #19.