ENH: Allow to specify extentions' CI
The 'master+DL release ' column in this repo's README links to an
extension's own CI runs on master. This used to assume Travis. Allow to
specify the CI used.

Introducing 'ci' and optionally 'ci-user' keywords for extensions.yaml.

Adjusted extensions.yaml to reflect the switch to AppVeyor for
datalad-hirni and datalad-ukbiobank.